requirements.txt
PyYaml
aligo